Navigator fix by shruti
