+++ Source/Navigators/AuthNavigator.js
@@ -7,16 +7,13 @@ import {
   createStackNavigator,
 } from 'react-navigation';
 import LoginScreen from '../Screens/LoginScreenNew2';
-import CompleteProfileNavigator from './CompleteProfileNavigator'
+
   const AuthNavigator = createStackNavigator({
     LoginScreen:{
       screen:LoginScreen
-    },
-    // CompleteProfileNavigator: {
-    //   screen: CompleteProfileNavigator
-    // }
-   },{
-    initialRouteName:'LoginScreen', 
+    }},
+    {
+    initialRouteName:'LoginScreen',
     headerMode:'none'
    });
