+++ Source/Screens/Profile.js
@@ -1,100 +1,119 @@
 import React, { Component } from 'react';
-import {
-  View,
-  StyleSheet,
-} from 'react-native';
-import {Avatar} from 'react-native-elements';
-import ProfileTab from '../Navigators/ProfileTab';
-//import ImagePicker from 'react-native-image-picker';
-var ImagePicker = require('react-native-image-picker');
-//Not in use right now but can be used in future for 
-export default class IntroductionPage extends Component {
-  constructor(props) {
-    super(props);
-    this.state={
-        avatarSource:'', 
-        firstName:'',
-        lastName:'',
-        sex:'M',
-        dateOfBirth:'',
-        address:'',
-        zipcode:'',
-        country:'',
-        city:'',
-        email:'',
-        phoneNumber:'',
-    }
-  }
+import { StyleSheet, View, Button } from 'react-native';
+import { ViewPager } from 'rn-viewpager';
 
-  componentDidMount() {
-      
-   }
+import StepIndicator from 'react-native-step-indicator';
 
-  onAvatarClick = ()=>{
-      var options = {
-          title: 'Select Avatar',
-          customButtons: [
-            {name: 'fb', title: 'Choose Photo from Facebook'},
-          ],
-          storageOptions: {
-            skipBackup: true,
-            path: 'images'
-          }
-        }
+import PersonalInfo from './PersonalInfo';
+import BillingInfo from './BillingInfo';
+import ProfessionalInfo from './ProfessionalInfo';
+import PrivacyPolicyScreen from './PrivacyPolicyScreen';
+import HomeStack from '../Navigators/HomeStack';
 
-      ImagePicker.showImagePicker(options, (response) => {
-      console.log('Response = ', response);
 
-      if (response.didCancel) {
-        console.log('User cancelled image picker');
-      }
-      else if (response.error) {
-        console.log('ImagePicker Error: ', response.error);
-      }
-      else if (response.customButton) {
-        console.log('User tapped custom button: ', response.customButton);
-      }
-      else {
-        let source = { uri: response.uri };
-
-        // You can also display the image using data:
-        // let source = { uri: 'data:image/jpeg;base64,' + response.data };
-
-        this.setState({
-          avatarSource: source
-        });
-      }
-    });
-  } 
-
-  render() {
-    return (
-        <View style = {styles.container}>
-          <Avatar
-                      large
-                      rounded
-                      source = {this.state.avatarSource}
-                      onPress={() => this.onAvatarClick()}
-                      activeOpacity={0.7}
-          />
-          <View style={{height:'100%',width:'100%',}}>        
-          <ProfileTab/>
-        </View>               
-
-        </View>
-        )
-  }
+const firstIndicatorStyles = {
+    stepIndicatorSize: 25,
+    currentStepIndicatorSize: 30,
+    separatorStrokeWidth: 3,
+    stepStrokeWidth: 0,
+    stepStrokeCurrentColor: 'orange',
+    currentStepStrokeWidth: 3,
+    separatorFinishedColor: 'orange',
+    separatorUnFinishedColor: '#ffc57f',
+    stepIndicatorFinishedColor: 'orange',
+    stepIndicatorUnFinishedColor: '#ffc57f',
+    stepIndicatorCurrentColor: '#ffffff',
+    stepIndicatorLabelFontSize: 15,
+    currentStepIndicatorLabelFontSize: 15,
+    stepIndicatorLabelCurrentColor: '#000000',
+    stepIndicatorLabelFinishedColor: '#ffffff',
+    stepIndicatorLabelUnFinishedColor: 'rgba(255,255,255,0.5)',
+    labelColor: '#666666',
+    labelSize: 12,
+    currentStepLabelColor: 'orange'
 }
 
-const styles= StyleSheet.create({
-  container: {
-    paddingTop: 40,
-    paddingBottom:32,
-    backgroundColor: '#F57F17',
-    // backgroundColor: '#D4cdb1',
-    justifyContent: 'center',
-    alignItems: 'center',
-    height:'100%'
-  }
-});
+export default class Profile extends Component {
 
+    constructor(props) {
+        super(props);
+        this.state = {
+            currentPage: 0,
+        }
+
+        console.log(props);
+    }
+
+    componentWillReceiveProps(nextProps, nextState) {
+        if (nextState.currentPage != this.state.currentPage) {
+            if (this.viewPager) {
+                this.viewPager.setPage(nextState.currentPage)
+            }
+        }
+    }
+
+    render() {
+        USER_EMAIL = this.props.navigation.getParam('USER_EMAIL', 'test email', 
+                                                        'USER_ID', 'NO_ID', 
+                                                        'TERMS', 'test terms');
+        console.log(USER_EMAIL);
+        var PAGES = [
+            <PersonalInfo
+                USER_EMAIL={() => USER_EMAIL}
+                USER_ID={() => USER_ID}
+                onNextPressed={() => this.viewPager.setPage(1)}
+            />,
+            <BillingInfo
+                USER_ID={() => USER_ID}
+                onNextPressed={() => { this.viewPager.setPage(2) }}
+                onPrevPressed={() => { this.viewPager.setPage(0) }}
+            />,
+            <PrivacyPolicyScreen
+                TERMS={() => TERMS}
+                onNextPressed={() => {
+                    this.props.navigation.navigate('thankyouScreen')
+                }}
+                onPrevPressed={() => { this.viewPager.setPage(1) }}
+            />
+        ];
+        return (
+            <View style={styles.container}>
+                <View style={styles.stepIndicator}>
+                    <StepIndicator
+                        customStyles={firstIndicatorStyles}
+                        currentPosition={this.state.currentPage}
+                        labels={["Profile", "Billing\nDetails", "Agreement"]}
+                        stepCount={3} />
+                </View>
+                <ViewPager
+                    style={{ flexGrow: 1 }}
+                    ref={(viewPager) => { this.viewPager = viewPager }}
+                    onPageSelected={(page) => { this.setState({ currentPage: page.position }) }}
+                    onPageScroll={() => { console.log("sbcsbcsdb") }}
+                    scrollEnabled = {false}
+                    horizontalScroll = {false}
+                >
+                    {PAGES.map((page) => this.renderViewPagerPage(page))}
+                </ViewPager>
+            </View>
+        );
+    }
+
+    renderViewPagerPage = (data) => {
+        return (
+            <View>
+                <View>{data}</View>
+            </View>
+        )
+    }
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#ffffff',
+    },
+    stepIndicator: {
+        marginVertical: 10,
+    },
+});
